requirements.txt
beautifulsoup4==4.12.3
breadability==0.1.20
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
docopt==0.6.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
idna==3.6
jieba3k==0.35.1
joblib==1.3.2
lxml==5.1.0
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
pillow==10.2.0
pycountry==23.12.11
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
sumy==0.11.0
tinysegmenter==0.3
tldextract==5.1.1
tqdm==4.66.2
urllib3==2.2.1
Wand==0.6.13
